requirements.txt
pandas
numpy
pyyaml
xgboost
scikit-learn